requirements.txt
asn1crypto==0.24.0
blinker==1.4
brotlipy==0.7.0
certifi==2018.11.29
cffi==1.11.5
chardet==3.0.4
click==6.7
cryptography==2.3.1
Django==2.1.5
Flask==1.0.2
get==2018.11.19
greenlet==0.4.15
h11==0.7.0
h2==3.0.1
hpack==3.0.0
hyperframe==5.1.0
idna==2.8
image==1.5.27
itsdangerous==1.1.0
Jinja2==2.10
kaitaistruct==0.8
ldap3==2.5.1
MarkupSafe==1.1.0
mitmproxy==4.0.4
msgpack==0.6.0
neovim==0.3.1
numpy==1.15.4
opencv-contrib-python==3.4.4.19
opencv-python==3.4.4.19
passlib==1.7.1
Pillow==5.4.1
post==2018.11.20
public==2018.11.20
pyasn1==0.4.4
pycparser==2.19
pynvim==0.3.1
pyOpenSSL==18.0.0
pyparsing==2.2.2
pyperclip==1.6.5
pytz==2018.7
query-string==2018.11.20
request==2018.11.20
requests==2.21.0
ruamel.yaml==0.15.81
six==1.12.0
sortedcontainers==2.0.5
tornado==5.1.1
urllib3==1.24.1
urwid==2.0.1
Werkzeug==0.14.1
wsproto==0.11.0
